main.py
import asyncio
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

async def main():
    """
    Main entry point for ProjectIntakeAgentTwo.
    Currently a placeholder to verify environment setup.
    """
    print("🚀 ProjectIntakeAgentTwo is starting...")
    
    # Check for critical environment variables
    openai_key = os.getenv("OPENAI_API_KEY")
    if not openai_key:
        print("⚠️  WARNING: OPENAI_API_KEY is not set in .env")
    else:
        print("✅ OPENAI_API_KEY found.")

    print("\nSystem Architecture: Parallel Orchestrator ('Whisper Engine')")
    print("Risk Level: Medium (Score 8)")
    print("Ready to initialize agents...")

if __name__ == "__main__":
    asyncio.run(main())
